Module TypedC.Pred

type unop =

| Extract of Units.In_bits.t * Units.In_bits.t
(*
start index, length
*)

type binop =

| Add

| Sub

| Mul

| Div

| And

| Or

| Mod

| Concat of Units.In_bits.t * Units.In_bits.t
(*
Size 1, size 2. First argument is the most significant
*)

type expr =

| Const of Z.t

| Val of value

| Self

| Unop of unop * expr

| Binop of binop * expr * expr

type cmpop =

| Equal

| NotEqual

| ULt

| SLt

| ULeq

| SLeq

| UGeq

| SGeq

| UGt

| SGt

type mutval = {

id : int;

mutable value : bool;

}

type t =

| True

| Cmp of cmpop * expr * expr

| And of t * t

| Mutval of mutval * t

Predicate on a structure field or array element, which is supposed to be true at all times.

val evaluate_mutval : mutval -> t -> t

given a mutval, evaluate the value of the associated predicate

val init_mutval : t -> t

val is_true : t -> bool

check if a predicate can be immediately reduced to true, this is not trivial since mutval exists

val true_ : t

returns the true predicate

val equal : t -> t -> bool

val pp_binop : Format.formatter -> binop -> unit

val pp_cmpop : Format.formatter -> cmpop -> unit

val pp : Format.formatter -> t -> unit

val array_length_to_expr : array_length -> expr

val conjunction : t -> t -> t

Conjoint two predicate. Ignores true (void) predicates